Extract broadcast banner methods into WindowBroadcastMixin
First step of decomposing the window.py god-object, following the existing
WindowActions mixin pattern. Move the 14 broadcast-banner methods verbatim
into sshpilot/window_broadcast.py as WindowBroadcastMixin and have MainWindow
inherit it. Pure code move — every method keeps its signature and self. state
access, so behavior is unchanged.

Place the mixin before WindowActions in the bases so its
on_broadcast_command_action wins, and delete the old dialog-based
on_broadcast_command_action from WindowActions: it was dead code (the
window.py body method already shadowed it via MRO), and leaving it would now
shadow the mixin instead.

Add tests/test_window_broadcast_mixin.py asserting every broadcast method
resolves to the mixin module (catches a stray body copy re-shadowing it), the
mixin precedes WindowActions in the MRO, and the dead duplicate stays gone.
