+++ sshpilot/actions.py
@@ -437,76 +437,6 @@ class WindowActions:
         if hasattr(self, 'apply_connection_sort_preset'):
             self.apply_connection_sort_preset(preset_id)
 
-    def on_broadcast_command_action(self, action, param=None):
-        """Handle broadcast command action - shows dialog to input command"""
-        try:
-            dialog = Adw.MessageDialog(
-                transient_for=self,
-                modal=True,
-                heading=_("Broadcast Command"),
-                body=_("Enter a command to send to all open SSH terminals:"),
-            )
-
-            entry = Gtk.Entry()
-            entry.set_placeholder_text(_("e.g., ls -la"))
-            entry.set_activates_default(True)
-            entry.set_hexpand(True)
-            dialog.set_extra_child(entry)
-
-            dialog.add_response('cancel', _('Cancel'))
-            dialog.add_response('send', _('Send'))
-            dialog.set_response_appearance('send', Adw.ResponseAppearance.SUGGESTED)
-            dialog.set_default_response('send')
-            dialog.set_close_response('cancel')
-
-            def on_response(dialog, response):
-                if response == 'send':
-                    command = entry.get_text().strip()
-                    if command:
-                        sent_count, failed_count = self.terminal_manager.broadcast_command(command)
-
-                        result_dialog = Adw.MessageDialog(
-                            transient_for=self,
-                            modal=True,
-                            heading=_("Command Sent"),
-                            body=_("Command sent to {} SSH terminals. {} failed.").format(sent_count, failed_count),
-                        )
-                        result_dialog.add_response('ok', _('OK'))
-                        result_dialog.present()
-                    else:
-                        error_dialog = Adw.MessageDialog(
-                            transient_for=self,
-                            modal=True,
-                            heading=_("Error"),
-                            body=_("Please enter a command to send."),
-                        )
-                        error_dialog.add_response('ok', _('OK'))
-                        error_dialog.present()
-                dialog.destroy()
-
-            dialog.connect('response', on_response)
-            dialog.present()
-
-            def focus_entry():
-                entry.grab_focus()
-                return False
-
-            GLib.idle_add(focus_entry)
-
-        except Exception as e:
-            logger.error(f"Failed to show broadcast command dialog: {e}")
-            try:
-                error_dialog = Adw.MessageDialog(
-                    transient_for=self,
-                    modal=True,
-                    heading=_("Error"),
-                    body=_("Failed to open broadcast command dialog: {}").format(str(e)),
-                )
-                error_dialog.add_response('ok', _('OK'))
-                error_dialog.present()
-            except Exception:
-                pass
-
     def on_edit_known_hosts_action(self, action, param=None):
         """Open the known hosts editor window."""
         try:
